Reporting Directory Reorganization #174
