+++ python/cli/discoveries_cli.py
@@ -16,15 +16,15 @@ from typing import List, Optional
 
 from python.framework.discoveries.data_coverage.data_coverage_report_cache import DataCoverageReportCache
 from python.framework.discoveries.discovery_cache_manager import DiscoveryCacheManager
-from python.framework.discoveries.market_analyzer import MarketAnalyzer
-from python.framework.discoveries.market_analyzer_cache import MarketAnalyzerCache
+from python.framework.discoveries.market_analyzer.market_analyzer import MarketAnalyzer
+from python.framework.discoveries.market_analyzer.market_analyzer_cache import MarketAnalyzerCache
 from python.framework.discoveries.discovery_cache import DiscoveryCache
 from python.framework.discoveries.extreme_move_scanner import ExtremeMoveScanner
 from python.framework.types.scenario_generator_types import (
     SymbolAnalysis,
 )
-from python.framework.reporting.market_report import print_analysis_report
-from python.framework.reporting.comparison_report import print_cross_instrument_ranking
+from python.framework.discoveries.market_analyzer.market_analyzer_report import print_analysis_report
+from python.framework.discoveries.market_analyzer.market_analyzer_comparison_report import print_cross_instrument_ranking
 from python.framework.logging.bootstrap_logger import get_global_logger
 from python.data_management.index.bars_index_manager import BarsIndexManager
