+++ python/framework/batch_reporting/broker_info_renderer.py
