+++ python/framework/batch_reporting/grid/console_grid_renderer.py
