+++ python/framework/data_preparation/broker_data_preparator.py
@@ -17,7 +17,7 @@ from python.configuration.market_config_manager import MarketConfigManager
 from python.framework.logging.scenario_logger import ScenarioLogger
 from python.framework.factory.broker_config_factory import BrokerConfigFactory
 from python.framework.trading_env.broker_config import BrokerConfig, BrokerType
-from python.framework.reporting.broker_info_renderer import BrokerInfoRenderer
+from python.framework.batch_reporting.broker_info_renderer import BrokerInfoRenderer
 from python.framework.types.scenario_set_types import BrokerScenarioInfo, SingleScenario
 from python.framework.types.validation_types import ValidationResult
