+++ python/framework/discoveries/market_analyzer/market_analyzer_cache.py
@@ -31,7 +31,7 @@ import pyarrow.parquet as pq
 
 from python.configuration.app_config_manager import AppConfigManager
 from python.data_management.index.bars_index_manager import BarsIndexManager
-from python.framework.discoveries.market_analyzer import MarketAnalyzer
+from python.framework.discoveries.market_analyzer.market_analyzer import MarketAnalyzer
 from python.framework.logging.abstract_logger import AbstractLogger
 from python.framework.logging.bootstrap_logger import get_global_logger
 from python.framework.types.market_config_types import MarketType
